More codebase butchering :
** PLANNER **
- The planner now use the same resource sim code as vessels, the dedicated code is gone.
- Some loss of functionality (greenhouse harvest specific code)
- The planner is now evaluated only once after a modification
- Added a planner "COMMS" section and related classes in the comms folder
- Editor comms can use an API handler, similar to the vessel comms handler
- Some UI tweaks here and there

** PARTMODULE HANDLING AND API **
- Refactored the IKerbalismModule interface into 2 separate interfaces, IBackgroundModule and IPlannerModule
- IBackgroundModule for background update handling
- IPlannerModule for planner update handling
- Removed the PlannerDelegate and BackgroundDelegate systems
- The BackgroundDelegate and ResourceUpdate API are still here, for now.
  But I think we should remove them, there is WIP replacement in PartModuleAPI for BackgroundDelegate, and ResourceUpdate is useless IMO.
- The PlannerDelegate API is replaced by the KerbalismPlannerUpdate() API method in PartModuleAPI
- Implemented IPlannerModule in all modules, moved emitter and passiveShield to IBackgroundModule

** OTHER CHANGES **
- Tweaked scrubber rate to account for kerbal individuality and rounding errors
- A few bugfixes to the habitat module
- Refactored and documented the resource sim API methods

+++ src/Kerbalism/System/Settings.cs
@@ -55,7 +55,7 @@ namespace KERBALISM
 
 			// signal
 			UnlinkedControl = Lib.ConfigEnum(cfg, "UnlinkedControl", UnlinkedCtrl.none);
-			DataRateMinimumBitsPerSecond = Lib.ConfigValue(cfg, "DataRateMinimumBitsPerSecond", 1.0f);
+			DataRateMinimumBitsPerSecond = Lib.ConfigValue(cfg, "DataRateMinimumBitsPerSecond", 1.0);
 			DataRateDampingExponent = Lib.ConfigValue(cfg, "DataRateDampingExponent", 6.0f);
 			DataRateDampingExponentRT = Lib.ConfigValue(cfg, "DataRateDampingExponentRT", 6.0f);
 			DataRateSurfaceExperiment = Lib.ConfigValue(cfg, "DataRateSurfaceExperiment", 0.3f);
@@ -167,7 +167,7 @@ namespace KERBALISM
 
 		// signal
 		public static UnlinkedCtrl UnlinkedControl;             // available control for unlinked vessels: 'none', 'limited' or 'full'
-		public static float DataRateMinimumBitsPerSecond;       // as long as there is a control connection, the science data rate will never go below this.
+		public static double DataRateMinimumBitsPerSecond;      // as long as there is a control connection, the science data rate will never go below this.
 		public static float DataRateDampingExponent;            // how much to damp data rate. stock is equivalent to 1, 6 gives nice values, RSS would use 4
 		public static float DataRateDampingExponentRT;          // same for RemoteTech
 		public static float DataRateSurfaceExperiment;          // transmission rate for surface experiments (Serenity DLC)
